# Conflicts:
#	src/main/java/BGU/Group13B/backend/Repositories/Interfaces/IProductRepository.java

+++ src/main/java/BGU/Group13B/backend/Repositories/Interfaces/IStoreScore.java
@@ -7,5 +7,5 @@ public interface IStoreScore {
 
     void modifyStoreScore(int userId,int storeId, int score);
 
-    int getStoreScore(int storeId);
+    float getStoreScore(int storeId);
 }
